+++ ts/features/euCovidCert/store/reducers/index.ts
@@ -1,15 +1,19 @@
-import * as pot from "italia-ts-commons/lib/pot";
 import { combineReducers } from "redux";
 import { Action } from "../../../../store/actions/types";
-import { IndexedById } from "../../../../store/helpers/indexer";
-import { EUCovidCertificateResponse } from "../../types/EUCovidCertificateResponse";
-import { euCovidCertByAuthCodeReducer } from "./byAuthCode";
+import {
+  euCovidCertByAuthCodeReducer,
+  EuCovidCertByIdState
+} from "./byAuthCode";
+import { currentReducer, EuCovidCertCurrentSelectedState } from "./current";
 
 export type EuCovidCertState = {
-  byAuthCode: IndexedById<pot.Pot<EUCovidCertificateResponse, Error>>;
+  byAuthCode: EuCovidCertByIdState;
+  current: EuCovidCertCurrentSelectedState | null;
 };
 
 export const euCovidCertReducer = combineReducers<EuCovidCertState, Action>({
   // save, using the AuthCode as key the the pot.Pot<EUCovidCertificateResponse, Error> response
-  byAuthCode: euCovidCertByAuthCodeReducer
+  byAuthCode: euCovidCertByAuthCodeReducer,
+  // The auth code and messageId that the user is viewing
+  current: currentReducer
 });