+++ ts/screens/messages/MessagesHomeScreen.tsx
@@ -10,6 +10,7 @@ import {
   NavigationScreenProps
 } from "react-navigation";
 import { connect } from "react-redux";
+import { IOStyles } from "../../components/core/variables/IOStyles";
 import MessagesArchive from "../../components/messages/MessagesArchive";
 import MessagesDeadlines from "../../components/messages/MessagesDeadlines";
 import MessagesInbox from "../../components/messages/MessagesInbox";
@@ -19,12 +20,13 @@ import { ScreenContentHeader } from "../../components/screens/ScreenContentHeade
 import TopScreenComponent from "../../components/screens/TopScreenComponent";
 import { MIN_CHARACTER_SEARCH_TEXT } from "../../components/search/SearchButton";
 import { SearchNoResultMessage } from "../../components/search/SearchNoResultMessage";
+import SectionStatusComponent from "../../components/SectionStatusComponent";
 import I18n from "../../i18n";
 import {
   loadMessages,
   setMessagesArchivedState
 } from "../../store/actions/messages";
-import { navigateToMessageDetailScreenAction } from "../../store/actions/navigation";
+import { navigateToMessageRouterScreen } from "../../store/actions/navigation";
 import { loadServiceDetail } from "../../store/actions/services";
 import { Dispatch } from "../../store/actions/types";
 import { lexicallyOrderedMessagesStateSelector } from "../../store/reducers/entities/messages";
@@ -42,8 +44,6 @@ import { makeFontStyleObject } from "../../theme/fonts";
 import customVariables from "../../theme/variables";
 import { HEADER_HEIGHT, MESSAGE_ICON_HEIGHT } from "../../utils/constants";
 import { setStatusBarColorAndBackground } from "../../utils/statusBar";
-import SectionStatusComponent from "../../components/SectionStatusComponent";
-import { IOStyles } from "../../components/core/variables/IOStyles";
 
 type Props = NavigationScreenProps &
   ReturnType<typeof mapStateToProps> &
@@ -340,7 +340,7 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
     dispatch(loadServiceDetail.request(serviceId));
   },
   navigateToMessageDetail: (messageId: string) =>
-    dispatch(navigateToMessageDetailScreenAction({ messageId })),
+    dispatch(navigateToMessageRouterScreen({ messageId })),
   updateMessagesArchivedState: (
     ids: ReadonlyArray<string>,
     archived: boolean